+++ src/cupertino-pane.ts
@@ -12,7 +12,6 @@ export class CupertinoPane {
   public screenHeightOffset: number;
   public preventDismissEvent: boolean = false;
   public preventedDismiss: boolean = false;
-  private iconCloseColor: string = '#7a7a7e';
   public rendered: boolean = false;
   
   public wrapperEl: HTMLDivElement;
@@ -73,8 +72,6 @@ export class CupertinoPane {
   private drawBaseElements() {
     // Parent 
     this.parentEl = this.settings.parentElement;
-    // Content user element
-    this.contentEl = this.el;
     
     // Wrapper
     this.wrapperEl = document.createElement('div');
@@ -82,9 +79,9 @@ export class CupertinoPane {
     if (this.settings.inverse) {
       this.wrapperEl.classList.add('inverse');
     }
-    if (this.el.className) {
-      this.wrapperEl.classList.add(this.el.className);
-    }
+    if (this.settings.cssClass) {
+      this.wrapperEl.className += ` ${this.settings.cssClass}`;
+    };
     let internalStyles: string = '';
     internalStyles += `
       .cupertino-pane-wrapper {
@@ -95,8 +92,9 @@ export class CupertinoPane {
       }
     `;
 
-    // Panel
+    // Panel (appying transform ASAP, avoid timeouts for animate:true)
     this.paneEl = document.createElement('div');
+    this.paneEl.style.transform = `translateY(${this.screenHeightOffset}px) translateZ(0px)`;
     this.paneEl.classList.add('pane');
     internalStyles += `
       .cupertino-pane-wrapper .pane {
@@ -108,19 +106,21 @@ export class CupertinoPane {
         right: 0px;
         margin-left: auto;
         margin-right: auto;
-        background: #ffffff;
-        box-shadow: 0 4px 16px rgba(0,0,0,.12);
+        background: var(--cupertino-pane-background, #ffffff);
+        color: var(--cupertino-pane-color, #333333);
+        box-shadow: var(--cupertino-pane-shadow, 0 4px 16px rgba(0,0,0,.12));
         will-change: transform;
         padding-top: 15px; 
-        border-radius: 20px 20px 0 0;
+        border-radius: var(--cupertino-pane-border-radius, 20px) 
+                       var(--cupertino-pane-border-radius, 20px) 
+                       0 0;
       }
       .cupertino-pane-wrapper.inverse .pane {
         padding-bottom: 15px; 
         border-radius: 0 0 20px 20px;
-      }
-      .cupertino-pane-wrapper.darkmode .pane {
-        background: #1c1c1d; 
-        color: #ffffff;
+        border-radius: 0 0
+                       var(--cupertino-pane-border-radius, 20px) 
+                       var(--cupertino-pane-border-radius, 20px);
       }
     `;
 
@@ -164,16 +164,13 @@ export class CupertinoPane {
       .cupertino-pane-wrapper .move {
         margin: 0 auto;
         height: 5px;
-        background: #c0c0c0;
+        background: var(--cupertino-pane-move-background, #c0c0c0);
         width: 36px;
         border-radius: 4px;
       }
-      .cupertino-pane-wrapper.darkmode .move {
-        background: #5a5a5e;
-      }
       .cupertino-pane-wrapper .draggable.over .move {
         width: 70px; 
-        background: rgba(225, 225, 225, 0.6);
+        background: var(--cupertino-pane-move-background, rgba(225, 225, 225, 0.6));
         ${Support.backdropFilter ? `
           backdrop-filter: saturate(180%) blur(20px);
           -webkit-backdrop-filter: saturate(180%) blur(20px);
@@ -195,16 +192,18 @@ export class CupertinoPane {
         width: 26px;
         height: 26px;
         position: absolute;
-        background: #ebebeb;
+        background: var(--cupertino-pane-destroy-button-background, #ebebeb);
         right: 20px;
         z-index: 14;
         border-radius: 100%;
         top: 16px;
       }
-      .cupertino-pane-wrapper.darkmode .destroy-button {
-        background: #424246;
-      }
     `;
+
+    // Content user element
+    this.contentEl = this.el;
+    this.contentEl.style.transition = `opacity ${this.settings.animationDuration}ms ${this.settings.animationType} 0s`;
+    this.contentEl.style.overflowX = 'hidden';
 
     // Backdrop
     internalStyles += `
@@ -256,21 +255,18 @@ export class CupertinoPane {
       this.drawBaseElements();
       await this.setBreakpoints();
 
-      // Necessary Inlines
+      // Necessary Inlines with breakpoints
       this.paneEl.style.height = `${this.getPaneHeight()}px`;
       if (this.settings.inverse) {
         this.paneEl.style.top = `-${this.breakpoints.bottomer}px`;
       }
-      this.paneEl.style.transform = `translateY(${this.screenHeightOffset}px) translateZ(0px)`;
-      this.contentEl.style.transition = `opacity ${this.settings.animationDuration}ms ${this.settings.animationType} 0s`;
-      this.contentEl.style.overflowX = 'hidden';
 
       // Show elements
       this.wrapperEl.style.display = 'block';
       this.contentEl.style.display = 'block';
       this.wrapperEl.classList.add('rendered');
       this.rendered = true;
-
+      
       if (this.settings.followerElement) {
         if (!<HTMLElement>document.querySelector(this.settings.followerElement)) {
           console.warn('Cupertino Pane: wrong follower element selector specified', this.settings.followerElement);
@@ -289,15 +285,11 @@ export class CupertinoPane {
         this.pushElement = <HTMLElement>document.querySelector(this.settings.pushElement);
       }
 
-      if (this.settings.darkMode) {
-        this.setDarkMode({enable: true});
-      }
-
       if ((this.settings.buttonClose && this.settings.buttonDestroy) && !this.settings.inverse) {
         this.paneEl.appendChild(this.destroyButtonEl);
         this.destroyButtonEl.addEventListener('click', (t) => this.destroy({animate:true, destroyButton: true}));
         this.destroyButtonEl.innerHTML = `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
-          <path fill="${this.iconCloseColor}" d="M278.6 256l68.2-68.2c6.2-6.2 6.2-16.4 0-22.6-6.2-6.2-16.4-6.2-22.6 0L256 233.4l-68.2-68.2c-6.2-6.2-16.4-6.2-22.6 0-3.1 3.1-4.7 7.2-4.7 11.3 0 4.1 1.6 8.2 4.7 11.3l68.2 68.2-68.2 68.2c-3.1 3.1-4.7 7.2-4.7 11.3 0 4.1 1.6 8.2 4.7 11.3 6.2 6.2 16.4 6.2 22.6 0l68.2-68.2 68.2 68.2c6.2 6.2 16.4 6.2 22.6 0 6.2-6.2 6.2-16.4 0-22.6L278.6 256z"/>
+          <path fill="var(--cupertino-pane-icon-close-color, #7a7a7e);" d="M278.6 256l68.2-68.2c6.2-6.2 6.2-16.4 0-22.6-6.2-6.2-16.4-6.2-22.6 0L256 233.4l-68.2-68.2c-6.2-6.2-16.4-6.2-22.6 0-3.1 3.1-4.7 7.2-4.7 11.3 0 4.1 1.6 8.2 4.7 11.3l68.2 68.2-68.2 68.2c-3.1 3.1-4.7 7.2-4.7 11.3 0 4.1 1.6 8.2 4.7 11.3 6.2 6.2 16.4 6.2 22.6 0l68.2-68.2 68.2 68.2c6.2 6.2 16.4 6.2 22.6 0 6.2-6.2 6.2-16.4 0-22.6L278.6 256z"/>
         </svg>`;
       }
 
@@ -578,16 +570,6 @@ export class CupertinoPane {
    */  
   public enableDrag(): void {
     this.disableDragEvents = false;
-  }
-
-  public setDarkMode(conf: {enable: boolean} = {enable: true}) {    
-    if (conf.enable) {
-      this.wrapperEl.classList.add('darkmode');
-      this.iconCloseColor = '#a8a7ae';  
-    } else {
-      this.wrapperEl.classList.remove('darkmode');
-      this.iconCloseColor = '#7a7a7e';
-    }
   }
 
   /**